@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 
 @Component({
   selector: 'app-login',
@@ -6,10 +7,31 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./login.component.css']
 })
 export class LoginComponent implements OnInit {
+  loginForm!: FormGroup;
 
-  constructor() { }
+  constructor(private fb: FormBuilder) { }
+
 
   ngOnInit(): void {
+    this.loginForm = this.fb.group({
+      email: ['', [Validators.required, Validators.email]],
+      pwd: [
+        '',
+        [
+          Validators.required,
+          Validators.minLength(6),
+          Validators.maxLength(15),
+        ],
+      ],
+    });
   }
 
+  login(){}
+
+  get Email(): FormControl {
+    return this.loginForm.get('email') as FormControl;
+  }
+  get PWD(): FormControl {
+    return this.loginForm.get('pwd') as FormControl;
+  }
 }
